[Bug 37278] Set X-Analytics header on XHR reqs
This patch sets a default 'X-Analytics' header on all XHR requests that
declare the request as having been generated by the WikipediaMetroTest
app. The header format is as follows:

   X-Analytics: app=WikipediaMetroTest; ver=1.2.3

(Where '1.2.3' is the app version.)

The X-Analytics header was implemented by the Zero team as a generic means of
annotating requests with analytics information. The header is included in the
log format that the Varnishes broadcast at the analytics cluster.